Separate calculation logics from EditModal
